@@ -2,16 +2,16 @@ import Highcharts from "highcharts/highstock";
 
 (function (HC) {
         /***
-        
+
         Each indicator requires mothods:
-        
+
         - getDefaultOptions()                           - returns object with default parameters, like period etc.
         - getValues(chart, series, options) - returns array of calculated values for indicator
         - getGraph(chart, series, options)  - returns path, or columns as SVG elemnts to add.
-                                                                                    Doesn't add to chart via renderer! 
-        
+                                                                                    Doesn't add to chart via renderer!
+
         ***/
-        
+
         /***
         indicators: [{
             id: 'series-id',
@@ -19,17 +19,17 @@ import Highcharts from "highcharts/highstock";
             params: {
                 period: 'x',
                 n: 'y'
-            },    
+            },
             styles: {
                 lineWidth: 'x',
                 strokeColor: 'y'
             }
         }]
-        
+
         ***/
-        
-        
-        
+
+
+
         var UNDEFINED,
                 Chart = HC.Chart,
                 Axis = HC.Axis,
@@ -39,17 +39,17 @@ import Highcharts from "highcharts/highstock";
         mathMax = Math.max,
         NUMBER = "number";
 
-        
+
         function error(name) {
                 if(window.console){
                         console.error(name);
                 }
         }
-        
+
         function defined(obj) {
             return obj !== UNDEFINED && obj !== null;
         }
-        
+
         function forceRedraw(s){
                 if(s.indicators) {
                         each(s.indicators, function(el, i) {
@@ -73,22 +73,22 @@ import Highcharts from "highcharts/highstock";
         HC.splat = function (obj) {
             return HC.isArray(obj) ? obj : [obj];
         };
-        
+
         HC.setOptions({
                 tooltip: {
                         followPointer: true
                 }
         });
 
-        
+
         /***
-        
+
         Wrappers:
-        
+
         ***/
         /*
         *   Upadte height when height is changed:
-        */ 
+        */
         HC.wrap(HC.Chart.prototype, 'setSize', function(p, w, h, a) {
                 p.call(this, w, h, false);
                 if(this.alignAxes) {
@@ -97,7 +97,7 @@ import Highcharts from "highcharts/highstock";
                         this.redraw(a);
                 }
         });
-        
+
         /*
         *  Remove corresponding indicators for series
         */
@@ -105,7 +105,7 @@ import Highcharts from "highcharts/highstock";
                 var tempIndics = [],
                         s = this,
                         tmpAxis, len, el;
-                        
+
                 if(s.indicators) {
                         len = s.indicators.length;
                         while(len--) { // #21
@@ -123,7 +123,7 @@ import Highcharts from "highcharts/highstock";
                         proceed.call(this, redraw, animation);
                 }
         });
-        
+
         /*
         *  Remove corresponding indicators for series
         */
@@ -137,10 +137,10 @@ import Highcharts from "highcharts/highstock";
                     };
                     s.indicators = null;
                 }
-                        
+
                 proceed.call(this, redraw, animation);
         });
-        
+
         /*
         *  Force redraw for indicator with new data
         */
@@ -151,7 +151,7 @@ import Highcharts from "highcharts/highstock";
                 } */
                 proceed.call(this, redraw, animation);
         });
-        
+
         /*
         *  Force redraw for indicator when new point is added
         */
@@ -159,20 +159,20 @@ import Highcharts from "highcharts/highstock";
                 forceRedraw(this);
                 proceed.call(this, options, redraw, shift, animation);
         });
-        
-        
+
+
         /*
-        *  Set visibility to true, but disable tooltip when needed. Required for forcing recalculation of values 
+        *  Set visibility to true, but disable tooltip when needed. Required for forcing recalculation of values
         */
         HC.wrap(HC.Series.prototype, 'setVisible', function(proceed, vis, redraw) {
                 var newVis = vis === UNDEFINED ? ( this.isVisible === UNDEFINED ? !this.visible : !this.isVisible) : vis,
                         showOrHide = newVis ? 'show' : 'hide',
                         series = this;
-                        
-                if(series.indicators) { 
+
+                if(series.indicators) {
                     series.isVisible = newVis;
                     series.hideInTooltip = !newVis;
-                    proceed.call(series, true, true); 
+                    proceed.call(series, true, true);
                     if(series.chart.legend && series.chart.legend.options.enabled) { // #20
                             series.chart.legend.colorizeItem(this, newVis);
                     }
@@ -188,7 +188,7 @@ import Highcharts from "highcharts/highstock";
                         proceed.call(series, newVis, true);
                 }
         });
-        
+
         /*
         *  Force redraw for indicator with new point options, like value
         */
@@ -196,7 +196,7 @@ import Highcharts from "highcharts/highstock";
                 forceRedraw(this.series);
                 proceed.call(this, options, redraw);
         });
-        
+
         /*
         *  Force redraw for indicator when one of points is removed
         */
@@ -204,7 +204,7 @@ import Highcharts from "highcharts/highstock";
                 forceRedraw(this.series);
                 proceed.call(this, options, redraw);
         });
-        
+
     /*
     *  Set flag for hasData when indicator has own axis
     */
@@ -216,7 +216,7 @@ import Highcharts from "highcharts/highstock";
                             indMax = -Infinity;
                         each(this.indicators, function(ind, i) {
                                 if(ind.visible) {
-                                        hasData = true; 
+                                        hasData = true;
                                         indMin = Math.min(indMin, ind.options.yAxisMin);
                                         indMax = Math.max(indMax, ind.options.yAxisMax);
                                 }
@@ -236,11 +236,11 @@ import Highcharts from "highcharts/highstock";
                         // case 1: axis doesn't have series
                         oldHasData = manageIndicators.call(this);
                         this.hasData = typeof this.hasData == "function" ? function() { return oldHasData; } : oldHasData; // version 4.1.6 vs 4.1.7
-                        
+
                         if((typeof this.hasData == "function" && this.hasData()) || (typeof this.hasData !== "function" && this.hasData)) { // version 4.1.6 vs 4.1.7
                                 this.setScale();
-                                this.setTickPositions(true);  
-        
+                                this.setTickPositions(true);
+
                                 this.chart.getMargins(); // #38
                                 HC.each(this.indicators, function(ind, e) {
                                     ind.drawGraph();
@@ -249,13 +249,13 @@ import Highcharts from "highcharts/highstock";
                 } else if(this.indicators) {
                         // case 2: right now all series are 'visible', so we need to check param: isVisible
                         var hasData = false;
-                        
+
                         each(this.series, function(serie, i) {
                                 if(serie.isVisible || (serie.isVisible === UNDEFINED && serie.visible)) {
-                                    hasData = true; 
+                                    hasData = true;
                                 }
                         });
-                        
+
                         if(!hasData) {
                                 hasData = manageIndicators.call(this);
                         } else {
@@ -264,12 +264,12 @@ import Highcharts from "highcharts/highstock";
                         }
                         if(hasData) {
                                 this.setScale();
-                                this.setTickPositions(true); 
-                        
+                                this.setTickPositions(true);
+
                                 this.chart.getMargins();
                         }
                         this.hasData = typeof this.hasData == "function" ? function() { return hasData } : hasData; // version 4.1.6 vs 4.1.7
-                        
+
                         HC.each(this.indicators, function(ind, e) {
                             ind.drawGraph();
                         });
@@ -277,7 +277,7 @@ import Highcharts from "highcharts/highstock";
             p.call(this);
     });
 
-    
+
     /*
         *       Tooltip formatter content
         */
@@ -312,7 +312,7 @@ import Highcharts from "highcharts/highstock";
         /*
             Tooltip pointFormat
         */
-        
+
         HC.wrap(HC.Point.prototype, 'tooltipFormatter', function (proceed, pointFormat) {
 
             // Insert options for valueDecimals, valuePrefix, and valueSuffix
@@ -329,7 +329,7 @@ import Highcharts from "highcharts/highstock";
                 indTooltip,
                 indPointFormat,
                 k;
-            
+
             // Loop over the point array map and replace unformatted values with sprintf formatting markup
             HC.each(series.pointArrayMap || ['y'], function (key) {
                 key = '{point.' + key; // without the closing bracket
@@ -340,7 +340,7 @@ import Highcharts from "highcharts/highstock";
             });
 
             if(indicators && indicators !== UNDEFINED && tooltipOptions.enabledIndicators) {
-                
+
                 // build the values of indicators
                 HC.each(indicators,function(ind,i) {
                     if(typeof(ind.values) === 'undefined' || ind.visible === false) {
@@ -388,13 +388,13 @@ import Highcharts from "highcharts/highstock";
         });
 
         /***
-        
+
         Add legend items:
-        
+
         ***/
-        
-        
-        /* 
+
+
+        /*
         * Add indicators to legend
         */
         HC.wrap(HC.Legend.prototype, 'getAllItems', function(p) {
@@ -409,8 +409,8 @@ import Highcharts from "highcharts/highstock";
                 }
                 return allItems;
         });
-        
-        
+
+
         /*
         * Render indicator
         */
@@ -429,7 +429,7 @@ import Highcharts from "highcharts/highstock";
 
         HC.wrap(HC.Point.prototype, 'getLabelConfig', function(proceed, point, mouseEvent) {
             var point = this;
-        
+
             return {
                 x: point.category,
                 y: point.y,
@@ -441,17 +441,17 @@ import Highcharts from "highcharts/highstock";
                 total: point.total || point.stackTotal
             };
         });
-        
+
         HC.wrap(HC.Point.prototype, 'init', function(proceed, series, options, x) {
             var point = this,
                 colors;
-        
+
             point.series = series;
             point.color = series.color; // #3445
             point.applyOptions(options, x);
             point.pointAttr = {};
             point.indicators = {};
-        
+
             if (series.options.colorByPoint) {
                 colors = series.options.colors || series.chart.options.colors;
                 point.color = point.color || colors[series.colorCounter++];
@@ -460,43 +460,43 @@ import Highcharts from "highcharts/highstock";
                     series.colorCounter = 0;
                 }
             }
-        
+
             series.chart.pointCount++;
             return point;
         });
-        
-        /* 
+
+        /*
         * When hovering legend item, use isVisible instead of visible property
-        */ 
+        */
         HC.wrap(HC.Legend.prototype, 'setItemEvents', function(p, item, legendItem, useHTML, itemStyle, itemHiddenStyle) {
                 p.call(this, item, legendItem, useHTML, itemStyle, itemHiddenStyle);
                 (useHTML ? legendItem : item.legendGroup).on('mouseout', function () {
-                        var style = item.isVisible === UNDEFINED ? 
+                        var style = item.isVisible === UNDEFINED ?
                                                 (item.visible ? itemStyle : itemHiddenStyle) : (item.isVisible ? itemStyle : itemHiddenStyle);
                         legendItem.css(style);
                         item.setState();
                 })
         });
-        
+
         /***
-        
+
         Indicator Class:
-        
+
         ***/
-                
+
         window.Indicator = function () {
             this.init.apply(this, arguments);
         };
-        
+
         Indicator.prototype = {
-            /* 
+            /*
             * Initialize the indicator
             */
             init: function (chart, options) {
                 // set default params, when not specified in params
                 if(!Indicator.prototype[options.type]) error("Indicator: " + options.type + " not found!");
                 options.params = merge({}, Indicator.prototype[options.type].getDefaultOptions(), options.params);
-    
+
                 this.chart = chart;
                 this.options = options;
                 this.series = chart.get(options.id);
@@ -508,7 +508,7 @@ import Highcharts from "highcharts/highstock";
                 }
                 this.series.indicators.push(this);
             },
-            
+
             /*
             * Render the indicator
             */
@@ -526,18 +526,18 @@ import Highcharts from "highcharts/highstock";
                         arrayValues,
                         extremes;
 
-                if(!indicator.visible) return;      
-                        
+                if(!indicator.visible) return;
+
                 if (!group) {
                         indicator.group = group = renderer.g().add(chart.indicators.group);
                 }
-                
+
                 if(!series) {
                         error('Series not found');
                         return false;
                 } else if(!graph) {
                         arrayValues = Indicator.prototype[options.type].getValues(chart, { points: [] }, options, [series.xData, series.yData]);
-                        if(!arrayValues) { //#6 - create dummy data 
+                        if(!arrayValues) { //#6 - create dummy data
                             arrayValues = {
                                 values: [[]],
                                 xData: [[]],
@@ -550,7 +550,7 @@ import Highcharts from "highcharts/highstock";
                         this.yData = arrayValues.yData;
                         this.groupPoints(series);
                         this.graph = graph = Indicator.prototype[options.type].getGraph(chart, series, options, this.values);
-                        
+
                         if(graph) {
                             var len = graph.length,
                                 i;
@@ -579,7 +579,7 @@ import Highcharts from "highcharts/highstock";
             },
 
             /*
-            * Redraw the indicator 
+            * Redraw the indicator
             */
             redraw: function () {
                 var options = this.options,
@@ -593,18 +593,18 @@ import Highcharts from "highcharts/highstock";
                         pointsBeyondExtremes,
                         arrayValues,
                         extremes;
-                        
+
                 if(!this.visible) {
                     // remove extremes
                     options.yAxisMax = null;
                     options.yAxisMin = null;
                     this.values = [[]];
-                    return;             
+                    return;
                 }
                 // only after series.setData() or series.addPoint() etc.
                 if(isDirtyData) {
                     arrayValues = Indicator.prototype[options.type].getValues(chart, { points: [] }, options, [series.xData, series.yData]);
-                    if(!arrayValues) { //#6 - create dummy data 
+                    if(!arrayValues) { //#6 - create dummy data
                             arrayValues = {
                                     values: [[]],
                                     xData: [[]],
@@ -617,8 +617,8 @@ import Highcharts from "highcharts/highstock";
                 }
                 // always check if points should be grouped, like after setExtremes() which doesn't change data
                 this.groupPoints(series);
-            },  
-            
+            },
+
             /*
             * Draw graph
             */
@@ -627,27 +627,27 @@ import Highcharts from "highcharts/highstock";
                     graph = this.graph,
                     len = graph.length,
                     i;
-                        
+
                 if(graph) {
                         for(i = 0; i < len ;i++) {
                                 graph[i].destroy();
                         }
                 }
                 ind.graph = Indicator.prototype[ind.options.type].getGraph(ind.chart, ind.series, ind.options, ind.values);
-        
+
                 if(ind.graph) {
                         ind.clipPath.attr({
                                 x: ind.options.Axis.left,
                                 y: ind.options.Axis.top,
                                 width: ind.options.Axis.width,
                                 height: ind.options.Axis.height
-                        });   
+                        });
                         for(i = 0; i < len ;i++) {
                             ind.graph[i].add(ind.group);
-                        } 
+                        }
                 }
             },
-            
+
             /*
             * Group points to allow calculation before extremes
             */
@@ -657,13 +657,13 @@ import Highcharts from "highcharts/highstock";
                         end,
                         length = HC.splat(this.options.params.period)[0], //#23 - don't use cropShoulded - it's broken since v1.3.6
                         minLen = Math.max(0, start - length - 1), // cropping starts from 0 at least
-                        maxLen = series.options.data.length,              
+                        maxLen = series.options.data.length,
                         groupedPoints = [],
                         currentData = [],
                         xAxis = series.xAxis,
                         groupIntervalFactor,
                         interval,
-                        ex, 
+                        ex,
                         xMin, xMax,
                         yData, groupPositions,
                         processedXData,
@@ -720,10 +720,10 @@ import Highcharts from "highcharts/highstock";
                     } else {
                             this.processedXData = this.xData.slice(minLen, maxLen); //copy default data
                             this.processedYData = yData = this.yData.slice(minLen, maxLen);
-                    } 
+                    }
                     // merge current points
                     HC.each(this.processedXData, function(p, i){
-                            if(HC.isArray(yData[i])) { 
+                            if(HC.isArray(yData[i])) {
                                     currentData.push( [p].concat(yData[i]) );
                             } else {
                                     currentData.push( [p, yData[i]] );
@@ -749,46 +749,46 @@ import Highcharts from "highcharts/highstock";
                             values = [[], [], [], []],
                             approximationFn = typeof approximation === 'function' ? approximation : HC.approximations[approximation],
                             i;
-                
+
                     // Start with the first point within the X axis range (#2696)
                     for (i = 0; i <= dataLength; i++) {
                         if (xData[i] >= groupPositions[0]) {
                             break;
                         }
                     }
-                    
+
                     for (; i <= dataLength; i++) {
                         // when a new group is entered, summarize and initiate the previous group
                         while ((groupPositions[1] !== UNDEFINED && xData[i] >= groupPositions[1]) ||
                                 i === dataLength) { // get the last group
-                
+
                             // get group x and y
                             pointX = groupPositions.shift();
                             groupedY = approximationFn.apply(0, values);
-                
+
                             // push the grouped data
                             if (groupedY !== UNDEFINED) {
                                 groupedXData.push(pointX);
                                 groupedYData.push(groupedY);
                             }
-                
+
                             // reset the aggregate arrays
                             values[0] = [];
                             values[1] = [];
                             values[2] = [];
                             values[3] = [];
-                
+
                             // don't loop beyond the last group
                             if (i === dataLength) {
                                 break;
                             }
                         }
-                
+
                         // break out
                         if (i === dataLength) {
                             break;
                         }
-            
+
                         pointY = yData[i];
                         if (pointY === null) {
                             values[0].hasNulls = true;
@@ -800,10 +800,10 @@ import Highcharts from "highcharts/highstock";
                             });
                         }
                     }
-                        
+
                     return [groupedXData, groupedYData];
             },
-            
+
             /*
             * Apply indicator's value to the grouped, corresponding points
             */
@@ -816,7 +816,7 @@ import Highcharts from "highcharts/highstock";
                         diff = pLen - vLen,
                         point,
                         i;
-                        
+
                 for(i = diff; i < pLen; i++){
                     point = points[i];
                     if(point) {
@@ -847,7 +847,7 @@ import Highcharts from "highcharts/highstock";
                         allItems = chart.indicators.allItems,
                         index = allItems.indexOf(indicator),
                         Axis = this.options.Axis;
-                
+
                 // remove from all indicators array
                 if (index > -1) {
                     allItems.splice(index, 1);
@@ -863,47 +863,47 @@ import Highcharts from "highcharts/highstock";
                 if(index > -1) {
                         Axis.indicators.splice(index, 1);
                 }
-                
+
                 if(indicator.legendGroup) {
                         indicator.legendGroup.destroy();
                         if(chart.legend && chart.legend.options.enabled) {
                                 chart.legend.render();
                         }
                 }
-                
+
                 //remove axis if that was the last one indicator
                 if(Axis && Axis.series.length === 0 && Axis.indicators && Axis.indicators.length === 0) {
                     Axis.remove();
                     this.options.Axis = UNDEFINED;
-                    chart.indicators.haveAxes --; // #18: decrement number of axes to be updated        
+                    chart.indicators.haveAxes --; // #18: decrement number of axes to be updated
                     if(chart.alignAxes) {
                             chart.updateHeightAxes(20, false, true);
                     }
                 }
-                
+
                 // remove group with graph
                 if (indicator.group) {
                     indicator.group.destroy();
                     indicator.group = null;
                 }
-                
+
                 // remove points from tooltip #29
                 each(indicator.series.points, function(p) {
-                    if(p && p.indicators && p.indicators[indicator.options.type]){ 
+                    if(p && p.indicators && p.indicators[indicator.options.type]){
                         delete p.indicators[indicator.options.type];
                     }
         });
                 indicator = null;
                 chart.redraw(redraw);
             },
-            
+
             /*
             * setState for indicator?
             */
             setState: function(state) {
-                
+
             },
-            
+
             /*
             * Hide or show indicator
             */
@@ -912,34 +912,34 @@ import Highcharts from "highcharts/highstock";
                         oldVis = indicator.visible,
                         newVis,
                         method;
-                
+
                 if(vis === UNDEFINED) {
                         newVis = oldVis ? false : true;
                         method = oldVis ? 'hide' : 'show';
                 } else {
                         newVis = vis;
                         method = vis ? 'show' : 'hide';
-                }   
-                
+                }
+
                 if (this.options.showInLegend) {
                         this.chart.legend.colorizeItem(this, newVis);
                 }
                 this.visible = newVis;
-                
+
                 indicator[method]();
                 indicator.redraw();
-                
+
                 // hide axis by resetting extremes
                 if(this.options.Axis) {
                         this.options.Axis.render();
                 }
-                
-            }, 
-            
+
+            },
+
             /*
             * Draw symbol in legend - should be simple line
-            */ 
-            
+            */
+
             drawLegendSymbol: function(legend) {
                     var options = this.options,
                             markerOptions = options.marker,
@@ -951,7 +951,7 @@ import Highcharts from "highcharts/highstock";
                             legendItemGroup = this.legendGroup,
                             verticalCenter = legend.baseline - Math.round(renderer.fontMetrics(legendOptions.itemStyle.fontSize, this.legendItem).b * 0.3),
                             attr;
-                
+
                     // Draw the line
                     attr = {
                         'stroke-width': options.lineWidth || 2
@@ -970,17 +970,17 @@ import Highcharts from "highcharts/highstock";
                     .attr(attr)
                     .add(legendItemGroup);
             },
-            
+
             /*
             * Update the indicator with a given options
             */
             update: function (options, redraw) {
                 merge(true, this.options, options);
-                
+
                 this.redraw(redraw);
                 this.options.Axis.render();
             },
-            
+
             /*
             * Hide the indicator
             */
@@ -988,7 +988,7 @@ import Highcharts from "highcharts/highstock";
                     this.group.hide();
                     this.visible = false;
             },
-            
+
             /*
             * Show the indicator
             */
@@ -997,9 +997,9 @@ import Highcharts from "highcharts/highstock";
                     this.visible = true;
             }
         };
-        
-        
-    
+
+
+
         // Add indicator methods to chart prototype
         extend(Chart.prototype, {
                 /*
@@ -1021,7 +1021,7 @@ import Highcharts from "highcharts/highstock";
                  */
                 redrawIndicators: function () {
                         var chart = this;
-                        
+
                         each(chart.indicators.allItems, function (indicator) {
                                     indicator.redraw();
                         });
@@ -1030,10 +1030,10 @@ import Highcharts from "highcharts/highstock";
                         each(chart.yAxis, function (axis) {
                                     axis.render();
                         });
-                        
+
                 },
                 /*
-                 * updates axes and returns new and normalized height for each of them. 
+                 * updates axes and returns new and normalized height for each of them.
                  */
                 updateHeightAxes: function(topDiff, add, afterRemove) {
                         var chart = this,
@@ -1045,26 +1045,26 @@ import Highcharts from "highcharts/highstock";
                 indexWithoutNav = 0,
                 newHeight,
                 top;
-                
-            // #18 - don't update axes when none of indicators have separate axis 
-            if(afterRemove !== true && (!chart.indicators || chart.indicators.haveAxes == 0 || chart.indicators.allItems.length === 0)) return;   
-                
+
+            // #18 - don't update axes when none of indicators have separate axis
+            if(afterRemove !== true && (!chart.indicators || chart.indicators.haveAxes == 0 || chart.indicators.allItems.length === 0)) return;
+
             // when we want to remove axis, e.g. after indicator remove
             // #17 - we need to consider navigator (disabled vs enabled) when calculating height in advance
-            if(!add && chart.options.navigator.enabled) { 
-                calcLen--; 
+            if(!add && chart.options.navigator.enabled) {
+                calcLen--;
             } else if(add && !chart.options.navigator.enabled){
                 calcLen++;
             }
-            
+
                         newHeight = (sum - (calcLen-1) * topDiff) / calcLen;
             //update all axis
             for (;i < len; i++) {
                 var yAxis = chYxis[i];
-                
+
                 if(yAxis.options.id !== 'navigator-y-axis') {
                         top = chart.plotTop + indexWithoutNav * (topDiff + newHeight);
-                        
+
                                         if(yAxis.top !== top || yAxis.height !== newHeight) {
                                                 chYxis[i].update({
                                                         top: top,
@@ -1072,27 +1072,27 @@ import Highcharts from "highcharts/highstock";
                                                 }, false);
                                         }
                         indexWithoutNav++;
-                } 
+                }
             }
             return newHeight;
                 }
         });
-        
+
         // Add yAxis as pane
         extend(Axis.prototype, {
-                /* 
-                 * When new indicator is added, sometimes we need new pane. 
+                /*
+                 * When new indicator is added, sometimes we need new pane.
                  * Note: It automatically scales all of other axes unless alignAxes is set to false.
                  */
                 addAxisPane: function(chart, userOptions) {
                         chart.indicators.haveAxes++;    // #18: increment number of axes
-                    
+
                         var topDiff = 20,
                                 height,
                                 yLen = chart.options.navigator.enabled ? chart.yAxis.length - 1 : chart.yAxis.length, // #17 - don't count navigator
                                 defaultOptions,
                               options;
-                                
+
                         if(chart.alignAxes) {
                             height = chart.updateHeightAxes(topDiff, true),
                             defaultOptions = {
@@ -1113,16 +1113,16 @@ import Highcharts from "highcharts/highstock";
                                         max: 100
                                 };
                         }
-                        
+
                         options = merge(defaultOptions,userOptions);
-                        
+
                         //add new axis
                         chart.preventIndicators = true;
                         chart.addAxis(options, false, true, false);
                         chart.preventIndicators = false;
                         return chart.yAxis.length - 1;
                 },
-                
+
                 minInArray: function(arr) {
                         return arr.reduce(function(min, arr) {
                                 return Math.min(min, arr[1]);
@@ -1134,7 +1134,7 @@ import Highcharts from "highcharts/highstock";
                         }, -Infinity);
                 }
         });
-        
+
         // Initialize on chart load
         Chart.prototype.callbacks.push(function (chart) {
         var options = chart.options.indicators,
@@ -1142,35 +1142,35 @@ import Highcharts from "highcharts/highstock";
                 i = 0,
                         group,
                         exportingFlag = true;
-                        
+
         group = chart.renderer.g("indicators");
         group.attr({
                 zIndex: 2
         });
         group.add();
-        
+
         if(!chart.indicators) chart.indicators = {};
-        
+
         // initialize empty array for indicators
         if(!chart.indicators.allItems) chart.indicators.allItems = [];
-        
-        
+
+
         // link chart object to indicators
         chart.indicators.chart = chart;
-        
+
         // link indicators group element to the chart
         chart.indicators.group = group;
-        
+
         // counter for axes #18
         chart.indicators.haveAxes = 0;
         chart.alignAxes = defined(chart.options.chart.alignAxes) ? chart.options.chart.alignAxes : true;
-        
+
         for(i = 0; i < optionsLen; i++) {
                 chart.addIndicator(options[i], false);
                 if((chart.get(options[i].id).data.length - 1) <= options[i].params.period) // SPLAT?
                     exportingFlag = false;
         }
-        
+
                  // update indicators after chart redraw
                 Highcharts.addEvent(chart, 'redraw', function () {
                         if(!chart.preventIndicators) {
@@ -1178,7 +1178,7 @@ import Highcharts from "highcharts/highstock";
                         }
                         chart.preventIndicators = false;
                 });
-                  
+
                 if(exportingFlag && optionsLen > 0 && chart.series && chart.series.length > 0) { // #16 & #27
                         chart.isDirtyLegend = true;
                       chart.series[0].isDirty = true;
@@ -1186,12 +1186,12 @@ import Highcharts from "highcharts/highstock";
                         chart.redraw(false);
                 }
         });
-        
+
         HC.approximations = {
             sum: function (arr) {
                 var len = arr.length,
                     ret;
-    
+
                 // 1. it consists of nulls exclusively
                 if (!len && arr.hasNulls) {
                     ret = null;
@@ -1204,19 +1204,19 @@ import Highcharts from "highcharts/highstock";
                 }
                 // 3. it has zero length, so just return undefined
                 // => doNothing()
-    
+
                 return ret;
             },
             average: function (arr) {
                 var len = arr.length,
                     ret = HC.approximations.sum(arr);
-    
+
                 // If we have a number, return it divided by the length. If not, return
                 // null or undefined based on what the sum method finds.
                 if (typeof ret === NUMBER && len) {
                     ret = ret / len;
                 }
-    
+
                 return ret;
             },
             open: function (arr) {
@@ -1237,7 +1237,7 @@ import Highcharts from "highcharts/highstock";
                 high = HC.approximations.high(high);
                 low = HC.approximations.low(low);
                 close = HC.approximations.close(close);
-    
+
                 if (typeof open === NUMBER || typeof high === NUMBER || typeof low === NUMBER || typeof close === NUMBER) {
                     return [open, high, low, close];
                 }
@@ -1246,7 +1246,7 @@ import Highcharts from "highcharts/highstock";
             range: function (low, high) {
                 low = HC.approximations.low(low);
                 high = HC.approximations.high(high);
-    
+
                 if (typeof low === NUMBER || typeof high === NUMBER) {
                     return [low, high];
                 }
