@@ -1,29 +1,29 @@
 import React, {Component, PropTypes} from "react";
 import ReactDOM from "react-dom";
-import alt from "alt-instance"
-import connectToStores from "alt/utils/connectToStores"
-import cname from "classnames"
+import alt from "alt-instance";
+import connectToStores from "alt/utils/connectToStores";
+import cname from "classnames";
 import notify from "actions/NotificationActions";
 import {PrivateKey, Address, Aes, PublicKey, hash} from "graphenejs-lib";
 import {Apis, ChainConfig} from "graphenejs-ws";
-import PrivateKeyStore from "stores/PrivateKeyStore"
-import WalletUnlockActions from "actions/WalletUnlockActions"
-import WalletCreate from "components/Wallet/WalletCreate"
-import LoadingIndicator from "components/LoadingIndicator"
+import PrivateKeyStore from "stores/PrivateKeyStore";
+import WalletUnlockActions from "actions/WalletUnlockActions";
+import WalletCreate from "components/Wallet/WalletCreate";
+import LoadingIndicator from "components/LoadingIndicator";
 import Translate from "react-translate-component";
 import counterpart from "counterpart";
 
 import BalanceClaimActive from "../Wallet/BalanceClaimActive";
-import BalanceClaimActiveActions from "actions/BalanceClaimActiveActions"
-import BalanceClaimAssetTotal from "components/Wallet/BalanceClaimAssetTotal"
+import BalanceClaimActiveActions from "actions/BalanceClaimActiveActions";
+import BalanceClaimAssetTotal from "components/Wallet/BalanceClaimAssetTotal";
 import WalletDb from "stores/WalletDb";
-import ImportKeysStore from "stores/ImportKeysStore"
+import ImportKeysStore from "stores/ImportKeysStore";
 
-import GenesisFilter from "chain/GenesisFilter"
+import GenesisFilter from "chain/GenesisFilter";
 
 require("./ImportKeys.scss");
 
-var import_keys_assert_checking = false
+let import_keys_assert_checking = false;
 
 @connectToStores
 export default class ImportKeys extends Component {
@@ -36,13 +36,13 @@ export default class ImportKeys extends Component {
     }
 
     static getStores() {
-        return [ImportKeysStore]
+        return [ImportKeysStore];
     }
 
     static getPropsFromStores() {
         return {
             importing: ImportKeysStore.getState().importing
-        }
+        };
     }
 
     static defaultProps = {
@@ -71,48 +71,48 @@ export default class ImportKeys extends Component {
     }
 
     reset(e, keep_file_name) {
-        if(e) e.preventDefault()
-        var state = this._getInitialState(keep_file_name);
+        if(e) e.preventDefault();
+        let state = this._getInitialState(keep_file_name);
         this.setState(state, ()=> this.updateOnChange());
     }
 
     onWif(event) {
         event.preventDefault();
-        var value = this.refs.wifInput.value;
-        this.addByPattern(value)
+        let value = this.refs.wifInput.value;
+        this.addByPattern(value);
     }
 
     onCancel(e) {
-        if(e) e.preventDefault()
+        if(e) e.preventDefault();
         this.setState(this._getInitialState());
     }
 
     updateOnChange() {
-        BalanceClaimActiveActions.setPubkeys(Object.keys(this.state.imported_keys_public))
+        BalanceClaimActiveActions.setPubkeys(Object.keys(this.state.imported_keys_public));
     }
 
     getImportAccountKeyCount(keys_to_account) {
-        var account_keycount = {}
-        var found = false
+        let account_keycount = {};
+        let found = false;
         for(let key in keys_to_account)
             for(let account_name of keys_to_account[key].account_names) {
                 account_keycount[account_name] =
-                    (account_keycount[account_name] || 0) + 1
-                found = true
+                    (account_keycount[account_name] || 0) + 1;
+                found = true;
             }
-        return found ? account_keycount : null
+        return found ? account_keycount : null;
     }
 
     upload(evt) {
-        this.reset(null, true)
-        var file = evt.target.files[0]
-        var reader = new FileReader()
+        this.reset(null, true);
+        let file = evt.target.files[0];
+        let reader = new FileReader();
         reader.onload = evt => {
-            var contents = evt.target.result
+            let contents = evt.target.result;
             try {
-                var json_contents
+                let json_contents;
                 try {
-                    json_contents = JSON.parse(contents)
+                    json_contents = JSON.parse(contents);
                     // This is the only chance to encounter a large file,
                     // try this format first.
                     this._parseImportKeyUpload(json_contents, file.name, update_state => {
@@ -120,111 +120,111 @@ export default class ImportKeys extends Component {
                         this.setState(update_state, ()=> {
                             if( update_state.genesis_filter_finished ) {
                                 // try empty password, also display "Enter import file password"
-                                this._passwordCheck()
+                                this._passwordCheck();
                             }
-                        })
-                    })
+                        });
+                    });
                 } catch(e) {
                     //DEBUG console.log("... _parseImportKeyUpload",e)
                     try {
-                        if( ! json_contents) file.name + " is an unrecognized format"
-                        this._parseWalletJson(json_contents)
+                        if( ! json_contents) file.name + " is an unrecognized format";
+                        this._parseWalletJson(json_contents);
                     } catch(ee) {
                         if( ! this.addByPattern(contents))
-                            throw ee
+                            throw ee;
                     }
                     // try empty password, also display "Enter import file password"
-                    this._passwordCheck()
+                    this._passwordCheck();
                 }
             } catch(message) {
-                console.error("... ImportKeys upload error", message)
-                this.setState({import_file_message: message})
+                console.error("... ImportKeys upload error", message);
+                this.setState({import_file_message: message});
             }
-        }
-        reader.readAsText(file)
+        };
+        reader.readAsText(file);
     }
 
     /** BTS 1.0 client wallet_export_keys format. */
     _parseImportKeyUpload(json_contents, file_name, update_state) {
-        var password_checksum, unfiltered_account_keys
+        let password_checksum, unfiltered_account_keys;
         try {
-            password_checksum = json_contents.password_checksum
+            password_checksum = json_contents.password_checksum;
             if( ! password_checksum)
-                throw file_name + " is an unrecognized format"
+                throw file_name + " is an unrecognized format";
 
             if( ! Array.isArray(json_contents.account_keys))
-                throw file_name + " is an unrecognized format"
+                throw file_name + " is an unrecognized format";
 
-            unfiltered_account_keys = json_contents.account_keys
+            unfiltered_account_keys = json_contents.account_keys;
 
-        } catch(e) { throw e.message || e }
+        } catch(e) { throw e.message || e; }
 
         // BTS 1.0 wallets may have a lot of generated but unused keys or spent TITAN addresses making
         // wallets so large it is was not possible to use the JavaScript wallets with them.
 
-        var genesis_filter = new GenesisFilter
+        let genesis_filter = new GenesisFilter;
         if( ! genesis_filter.isAvailable() ) {
             update_state({ password_checksum, account_keys: unfiltered_account_keys,
-                genesis_filter_finished: true, genesis_filtering: false })
-            return
+                genesis_filter_finished: true, genesis_filtering: false });
+            return;
         }
         this.setState({ genesis_filter_initalizing: true }, ()=>// setTimeout(()=>
             genesis_filter.init(()=> {
-                var filter_status = this.state.genesis_filter_status
+                let filter_status = this.state.genesis_filter_status;
 
                 // FF < version 41 does not support worker threads internals (like blob urls)
-                // var GenesisFilterWorker = require("worker!workers/GenesisFilterWorker")
-                // var worker = new GenesisFilterWorker
+                // let GenesisFilterWorker = require("worker!workers/GenesisFilterWorker")
+                // let worker = new GenesisFilterWorker
                 // worker.postMessage({
                 //     account_keys: unfiltered_account_keys,
                 //     bloom_filter: genesis_filter.bloom_filter
                 // })
                 // worker.onmessage = event => { try {
-                //     var { status, account_keys } = event.data
+                //     let { status, account_keys } = event.data
                 //     // ...
                 // } catch( e ) { console.error('GenesisFilterWorker', e) }}
 
-                var account_keys = unfiltered_account_keys
+                let account_keys = unfiltered_account_keys;
                 genesis_filter.filter( account_keys, status => {
                     //console.log("import filter", status)
                     if( status.error === "missing_public_keys" ) {
-                        console.error("un-released format, just for testing")
+                        console.error("un-released format, just for testing");
                         update_state({ password_checksum, account_keys: unfiltered_account_keys,
-                            genesis_filter_finished: true, genesis_filtering: false })
-                        return
+                            genesis_filter_finished: true, genesis_filtering: false });
+                        return;
                     }
                     if( status.success ) {
-                        // var { account_keys } = event.data // if using worker thread
+                        // let { account_keys } = event.data // if using worker thread
                         update_state({ password_checksum, account_keys,
-                            genesis_filter_finished: true, genesis_filtering: false })
-                        return
+                            genesis_filter_finished: true, genesis_filtering: false });
+                        return;
                     }
                     if( status.initalizing !== undefined ) {
-                        update_state({ genesis_filter_initalizing: status.initalizing, genesis_filtering: true })
-                        return
+                        update_state({ genesis_filter_initalizing: status.initalizing, genesis_filtering: true });
+                        return;
                     }
                     if( status.importing === undefined ) {
                         // programmer error
-                        console.error('unknown status', status)
-                        return
+                        console.error("unknown status", status);
+                        return;
                     }
                     if( ! filter_status.length )
                         // first account
-                        filter_status.push( status )
+                        filter_status.push( status );
                     else {
-                        var last_account_name = filter_status[filter_status.length - 1].account_name
+                        let last_account_name = filter_status[filter_status.length - 1].account_name;
                         if( last_account_name === status.account_name )
                             // update same account
-                            filter_status[filter_status.length - 1] = status
+                            filter_status[filter_status.length - 1] = status;
                         else
                             // new account
-                            filter_status.push( status )
+                            filter_status.push( status );
                     }
-                    update_state({ genesis_filter_status: filter_status })
-                })
+                    update_state({ genesis_filter_status: filter_status });
+                });
             })
         //, 100)
-        )
+    );
     }
 
     /**
@@ -239,26 +239,26 @@ export default class ImportKeys extends Component {
 
     */
     _parseWalletJson(json_contents) {
-        var password_checksum
-        var encrypted_brainkey
-        var address_to_enckeys = {}
-        var account_addresses = {}
+        let password_checksum;
+        let encrypted_brainkey;
+        let address_to_enckeys = {};
+        let account_addresses = {};
 
-        var savePubkeyAccount = function (pubkey, account_name) {
+        let savePubkeyAccount = function (pubkey, account_name) {
             //replace BTS with GPH
-            pubkey = ChainConfig.address_prefix + pubkey.substring(3)
-            var address = PublicKey.fromPublicKeyString(pubkey).toAddressString()
-            var addresses = account_addresses[account_name] || []
-            address = "BTS" + address.substring(3)
+            pubkey = ChainConfig.address_prefix + pubkey.substring(3);
+            let address = PublicKey.fromPublicKeyString(pubkey).toAddressString();
+            let addresses = account_addresses[account_name] || [];
+            address = "BTS" + address.substring(3);
             //DEBUG console.log("... address",address,account_name)
-            addresses.push(address)
-            account_addresses[account_name] = addresses
-        }
+            addresses.push(address);
+            account_addresses[account_name] = addresses;
+        };
 
         try {
             if(! Array.isArray(json_contents)) {
                 //DEBUG console.log('... json_contents',json_contents)
-                throw new Error("Invalid wallet format")
+                throw new Error("Invalid wallet format");
             }
             for(let element of json_contents) {
 
@@ -266,65 +266,65 @@ export default class ImportKeys extends Component {
                     element.data.account_address &&
                     element.data.encrypted_private_key
                 ) {
-                    var address = element.data.account_address
-                    var enckeys = address_to_enckeys[address] || []
-                    enckeys.push(element.data.encrypted_private_key)
+                    let address = element.data.account_address;
+                    let enckeys = address_to_enckeys[address] || [];
+                    enckeys.push(element.data.encrypted_private_key);
                     //DEBUG console.log("... address",address,enckeys)
-                    address_to_enckeys[address] = enckeys
-                    continue
+                    address_to_enckeys[address] = enckeys;
+                    continue;
                 }
 
                 if( "account_record_type" == element.type) {
-                    var account_name = element.data.name
-                    savePubkeyAccount(element.data.owner_key, account_name)
+                    let account_name = element.data.name;
+                    savePubkeyAccount(element.data.owner_key, account_name);
                     for(let history of element.data.active_key_history) {
-                        savePubkeyAccount(history[1], account_name)
+                        savePubkeyAccount(history[1], account_name);
                     }
-                    continue
+                    continue;
                 }
 
                 if ( "property_record_type" == element.type &&
                     "encrypted_brainkey" == element.data.key
                 ) {
-                    encrypted_brainkey = element.data.value
-                    continue
+                    encrypted_brainkey = element.data.value;
+                    continue;
                 }
 
                 if( "master_key_record_type" == element.type) {
                     if( ! element.data)
-                        throw file.name + " invalid master_key_record record"
+                        throw file.name + " invalid master_key_record record";
 
                     if( ! element.data.checksum)
-                        throw file.name + " is missing master_key_record checksum"
+                        throw file.name + " is missing master_key_record checksum";
 
-                    password_checksum = element.data.checksum
+                    password_checksum = element.data.checksum;
                 }
 
             }
             if( ! encrypted_brainkey)
-                throw "Please use a BTS 1.0 wallet_export_keys file instead"
+                throw "Please use a BTS 1.0 wallet_export_keys file instead";
 
             if( ! password_checksum)
-                throw file.name + " is missing password_checksum"
+                throw file.name + " is missing password_checksum";
 
             if( ! enckeys.length)
-                throw file.name + " does not contain any private keys"
+                throw file.name + " does not contain any private keys";
 
-        } catch(e) { throw e.message || e }
+        } catch(e) { throw e.message || e; }
 
-        var account_keys = []
+        let account_keys = [];
         for(let account_name in account_addresses) {
-            var encrypted_private_keys = []
+            let encrypted_private_keys = [];
             for(let address of account_addresses[account_name]) {
-                var enckeys = address_to_enckeys[address]
-                if( ! enckeys) continue
+                let enckeys = address_to_enckeys[address];
+                if( ! enckeys) continue;
                 for(let enckey of enckeys)
-                    encrypted_private_keys.push(enckey)
+                    encrypted_private_keys.push(enckey);
             }
             account_keys.push({
                 account_name,
                 encrypted_private_keys
-            })
+            });
         }
         // We could prompt for this brain key instead on first use.  The user
         // may already have a brainkey at this point so with a single brainkey
@@ -333,18 +333,18 @@ export default class ImportKeys extends Component {
             password_checksum,
             account_keys
             //encrypted_brainkey
-        })
+        });
     }
 
     _passwordCheck(evt) {
         if (evt && "preventDefault" in evt) {
             evt.preventDefault();
         }
-        var pwNode = this.refs.password;
+        let pwNode = this.refs.password;
         // if(pwNode) pwNode.focus()
-        var password = pwNode ? pwNode.value : ""
-        var checksum = this.state.password_checksum
-        var new_checksum = hash.sha512(hash.sha512(password)).toString("hex")
+        let password = pwNode ? pwNode.value : "";
+        let checksum = this.state.password_checksum;
+        let new_checksum = hash.sha512(hash.sha512(password)).toString("hex");
         if(checksum != new_checksum) {
             return this.setState({
                 no_file: false,
@@ -356,82 +356,82 @@ export default class ImportKeys extends Component {
             no_file: false,
             reset_password: Date.now(),
             import_password_message: counterpart.translate("wallet.import_pass_match")
-        }, ()=> this._decryptPrivateKeys(password))
+        }, ()=> this._decryptPrivateKeys(password));
         // setTimeout(, 250)
     }
 
     _decryptPrivateKeys(password) {
-        var password_aes = Aes.fromSeed(password)
-        var format_error1_once = true
+        let password_aes = Aes.fromSeed(password);
+        let format_error1_once = true;
         for(let account of this.state.account_keys) {
             if(! account.encrypted_private_keys) {
-                var error = `Account ${account.account_name} missing encrypted_private_keys`
-                console.error(error)
+                let error = `Account ${account.account_name} missing encrypted_private_keys`;
+                console.error(error);
                 if(format_error1_once) {
-                    notify.error(error)
-                    format_error1_once = false
+                    notify.error(error);
+                    format_error1_once = false;
                 }
-                continue
+                continue;
             }
-            var account_name = account.account_name.trim()
-            var same_prefix_regex = new RegExp("^" + ChainConfig.address_prefix)
+            let account_name = account.account_name.trim();
+            let same_prefix_regex = new RegExp("^" + ChainConfig.address_prefix);
             for(let i = 0; i < account.encrypted_private_keys.length; i++) {
-                let encrypted_private = account.encrypted_private_keys[i]
+                let encrypted_private = account.encrypted_private_keys[i];
                 let public_key_string = account.public_keys ?
-                    account.public_keys[i] : null // performance gain
+                    account.public_keys[i] : null; // performance gain
 
                 try {
-                    var private_plainhex = password_aes.decryptHex(encrypted_private)
+                    let private_plainhex = password_aes.decryptHex(encrypted_private);
                     if(import_keys_assert_checking && public_key_string) {
-                        var private_key = PrivateKey.fromHex( private_plainhex )
-                        var pub = private_key.toPublicKey() // S L O W
-                        var addy = pub.toAddressString()
-                        var pubby = pub.toPublicKeyString()
-                        var error = ""
+                        let private_key = PrivateKey.fromHex( private_plainhex );
+                        let pub = private_key.toPublicKey(); // S L O W
+                        let addy = pub.toAddressString();
+                        let pubby = pub.toPublicKeyString();
+                        let error = "";
 
                         let address_string = account.addresses ?
-                            account.addresses[i] : null // assert checking
+                            account.addresses[i] : null; // assert checking
 
                         if(address_string && addy.substring(3) != address_string.substring(3))
-                            error = "address imported " + address_string + " but calculated " + addy + ". "
+                            error = "address imported " + address_string + " but calculated " + addy + ". ";
 
                         if(pubby.substring(3) != public_key_string.substring(3))
-                            error += "public key imported " + public_key_string + " but calculated " + pubby
+                            error += "public key imported " + public_key_string + " but calculated " + pubby;
 
                         if(error != "")
-                            console.log("ERROR Miss-match key",error)
+                            console.log("ERROR Miss-match key",error);
                     }
 
                     if( ! public_key_string) {
-                        var private_key = PrivateKey.fromHex( private_plainhex )
-                        var public_key = private_key.toPublicKey()// S L O W
-                        public_key_string = public_key.toPublicKeyString()
+                        let private_key = PrivateKey.fromHex( private_plainhex );
+                        let public_key = private_key.toPublicKey(); // S L O W
+                        public_key_string = public_key.toPublicKeyString();
                     } else {
                         if( ! same_prefix_regex.test(public_key_string))
                             // This was creating a unresponsive chrome browser
                             // but after the results were shown.  It was probably
                             // caused by garbage collection.
                             public_key_string = ChainConfig.address_prefix +
-                                public_key_string.substring(3)
+                                public_key_string.substring(3);
                     }
-                    this.state.imported_keys_public[public_key_string] = true
-                    var {account_names} = this.state.keys_to_account[private_plainhex] ||
-                        {account_names: []}
-                    var dup = false
+                    this.state.imported_keys_public[public_key_string] = true;
+                    let {account_names} = this.state.keys_to_account[private_plainhex] ||
+                        {account_names: []};
+                    let dup = false;
                     for(let _name of account_names)
                         if(_name == account_name)
-                            dup = true
-                    if(dup) continue
-                    account_names.push(account_name)
-                    this.state.keys_to_account[private_plainhex] = {account_names, public_key_string}
+                            dup = true;
+                    if(dup) continue;
+                    account_names.push(account_name);
+                    this.state.keys_to_account[private_plainhex] = {account_names, public_key_string};
                 } catch(e) {
-                    console.log(e, e.stack)
-                    var message = e.message || e
-                    notify.error(`Account ${account_name} had a private key import error: `+message)
+                    console.log(e, e.stack);
+                    let message = e.message || e;
+                    notify.error(`Account ${account_name} had a private key import error: `+message);
                 }
             }
         }
-        //var enc_brainkey = this.state.encrypted_brainkey
+        //let enc_brainkey = this.state.encrypted_brainkey
         //if(enc_brainkey){
         //    this.setState({
         //        brainkey: password_aes.decryptHexToText(enc_brainkey)
@@ -441,49 +441,49 @@ export default class ImportKeys extends Component {
             import_file_message: null,
             import_password_message: null,
             password_checksum: null
-        }, ()=> this.updateOnChange())
+        }, ()=> this.updateOnChange());
     }
 
     _saveImport(e) {
-        e.preventDefault()
-        var keys = PrivateKeyStore.getState().keys
-        var dups = {}
+        e.preventDefault();
+        let keys = PrivateKeyStore.getState().keys;
+        let dups = {};
         for(let public_key_string in this.state.imported_keys_public) {
-            if( ! keys.has(public_key_string) ) continue
-            delete this.state.imported_keys_public[public_key_string]
-            dups[public_key_string] = true
+            if( ! keys.has(public_key_string) ) continue;
+            delete this.state.imported_keys_public[public_key_string];
+            dups[public_key_string] = true;
         }
         if( Object.keys(this.state.imported_keys_public).length === 0 ) {
-            notify.error("This wallet has already been imported")
-            return
+            notify.error("This wallet has already been imported");
+            return;
         }
-        var keys_to_account = this.state.keys_to_account
+        let keys_to_account = this.state.keys_to_account;
         for(let private_plainhex of Object.keys(keys_to_account)) {
-            var {account_names, public_key_string} = keys_to_account[private_plainhex]
-            if( dups[public_key_string] ) delete keys_to_account[private_plainhex]
+            let {account_names, public_key_string} = keys_to_account[private_plainhex];
+            if( dups[public_key_string] ) delete keys_to_account[private_plainhex];
         }
         WalletUnlockActions.unlock().then(()=> {
-            ImportKeysStore.importing(true)
+            ImportKeysStore.importing(true);
             // show the loading indicator
-            setTimeout(()=> this.saveImport(), 200)
-        })
+            setTimeout(()=> this.saveImport(), 200);
+        });
     }
 
     saveImport() {
-        var keys_to_account = this.state.keys_to_account
-        var private_key_objs = []
+        let keys_to_account = this.state.keys_to_account;
+        let private_key_objs = [];
         for(let private_plainhex of Object.keys(keys_to_account)) {
-            var {account_names, public_key_string} = keys_to_account[private_plainhex]
+            let {account_names, public_key_string} = keys_to_account[private_plainhex];
             private_key_objs.push({
                 private_plainhex,
                 import_account_names: account_names,
                 public_key_string
-            })
+            });
         }
-        this.reset()
+        this.reset();
         WalletDb.importKeysWorker( private_key_objs ).then( result => {
-            ImportKeysStore.importing(false)
-            var import_count = private_key_objs.length
+            ImportKeysStore.importing(false);
+            let import_count = private_key_objs.length;
 
             notify.success(counterpart.translate("wallet.import_key_success", {count: import_count}));
             this.setState({
@@ -491,41 +491,41 @@ export default class ImportKeys extends Component {
             });
             // this.onCancel() // back to claim balances
         }).catch( error => {
-            console.log("error:", error)
-            ImportKeysStore.importing(false)
-            var message = error
-            try { message = error.target.error.message } catch (e){}
-            notify.error(`Key import error: ${message}`)
-        })
+            console.log("error:", error);
+            ImportKeysStore.importing(false);
+            let message = error;
+            try { message = error.target.error.message; } catch (e){}
+            notify.error(`Key import error: ${message}`);
+        });
     }
 
     addByPattern(contents) {
         if( ! contents)
-            return false
+            return false;
 
-        var count = 0, invalid_count = 0
-        var wif_regex = /5[HJK][1-9A-Za-z]{49}/g
+        let count = 0, invalid_count = 0;
+        let wif_regex = /5[HJK][1-9A-Za-z]{49}/g;
         for(let wif of contents.match(wif_regex) || [] ) {
             try {
-                var private_key = PrivateKey.fromWif(wif) //could throw and error
-                var private_plainhex = private_key.toBuffer().toString('hex')
-                var public_key = private_key.toPublicKey() // S L O W
-                var public_key_string = public_key.toPublicKeyString()
-                this.state.imported_keys_public[public_key_string] = true
+                let private_key = PrivateKey.fromWif(wif); //could throw and error
+                let private_plainhex = private_key.toBuffer().toString("hex");
+                let public_key = private_key.toPublicKey(); // S L O W
+                let public_key_string = public_key.toPublicKeyString();
+                this.state.imported_keys_public[public_key_string] = true;
                 this.state.keys_to_account[private_plainhex] = {
-                    account_names: [], public_key_string}
-                count++
-            } catch(e) { invalid_count++ }
+                    account_names: [], public_key_string};
+                count++;
+            } catch(e) { invalid_count++; }
         }
         this.setState({
-            key_text_message: 'Found ' +
+            key_text_message: "Found " +
                 (!count ? "" : count + " valid") +
                 (!invalid_count ? "" : " and " + invalid_count + " invalid") +
                 " key" + ( count > 1 || invalid_count > 1 ? "s" : "") + "."
-        }, ()=> this.updateOnChange())
+        }, ()=> this.updateOnChange());
         // removes the message on the next render
-        this.state.key_text_message = null
-        return count
+        this.state.key_text_message = null;
+        return count;
     }
 
     // toggleImportType(type) {
@@ -553,27 +553,27 @@ export default class ImportKeys extends Component {
     }
 
     render() {
-        var {privateKey} = this.props;
-        var {keys_to_account} = this.state;
-        var key_count = Object.keys(keys_to_account).length
-        var account_keycount = this.getImportAccountKeyCount(keys_to_account)
+        let {privateKey} = this.props;
+        let {keys_to_account} = this.state;
+        let key_count = Object.keys(keys_to_account).length;
+        let account_keycount = this.getImportAccountKeyCount(keys_to_account);
 
         // Create wallet prior to the import keys (keeps layout clean)
-        if( ! WalletDb.getWallet()) return <WalletCreate importKeys={true} hideTitle={true}/>
+        if( ! WalletDb.getWallet()) return <WalletCreate importKeys={true} hideTitle={true}/>;
         if( this.props.importing ) {
             return <div>
                 <div className="center-content">
                     <LoadingIndicator type="circle"/>
                 </div>
-            </div>
+            </div>;
         }
 
-        var filtering = this.state.genesis_filtering
-        var was_filtered = !!this.state.genesis_filter_status.length && this.state.genesis_filter_finished
-        var account_rows = null;
+        let filtering = this.state.genesis_filtering;
+        let was_filtered = !!this.state.genesis_filter_status.length && this.state.genesis_filter_finished;
+        let account_rows = null;
 
         if(this.state.genesis_filter_status.length) {
-            account_rows = []
+            account_rows = [];
             for(let status of this.state.genesis_filter_status) {
                 if (status.count && status.total) {
                     account_rows.push(
@@ -584,15 +584,15 @@ export default class ImportKeys extends Component {
                                 : <span>{status.count}</span>
                             }</td>
                         </tr>
-                    )
+                    );
                 }
             }
         }
 
-        var import_ready = key_count !== 0
-        var password_placeholder = counterpart.translate("wallet.import_password");
+        let import_ready = key_count !== 0;
+        let password_placeholder = counterpart.translate("wallet.import_password");
 
-        if (import_ready) password_placeholder = ""
+        if (import_ready) password_placeholder = "";
 
         if( ! account_rows && account_keycount) {
             account_rows = [];
@@ -692,7 +692,7 @@ export default class ImportKeys extends Component {
                                             onChange={() => {if (this.state.import_password_message && this.state.import_password_message.length) {
                                                 this.setState({
                                                     import_password_message: null
-                                                })
+                                                });
                                             }}}
                                         />
                                         <p className="facolor-error">{this.state.import_password_message}</p>
@@ -752,7 +752,7 @@ export default class ImportKeys extends Component {
 
 class KeyCount extends Component {
     render() {
-        if( !this.props.key_count) return <span/>
-        return <span>Found {this.props.key_count} private keys</span>
+        if( !this.props.key_count) return <span/>;
+        return <span>Found {this.props.key_count} private keys</span>;
     }
 }
